compile warnings

+++ src/libespeak-ng/numbers.c
@@ -1683,7 +1683,7 @@ static int TranslateNumber_1(Translator *tr, char *word, char *ph_out, unsigned 
 		if (prev_thousands == 0) {
 			if ((decimal_point == 0) && (ordinal == 0)) {
 				// Look for special pronunciation for this number in isolation (LANG=kl)
-				sprintf(string, "_%dn", value);
+				sprintf(string, "_%dn", (int) value);
 				if (Lookup(tr, string, ph_out))
 					return 1;
 			}
